requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0.8
django-bootstrap4==2.2.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-extensions==3.0.2
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
future==0.18.2
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
kombu==4.6.11
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
xmltodict==0.12.0
